Add reusable Kraken balance util and integrate checks
